Organized Folder Structure & Navbar With Chakra
