+++ src/App.css
@@ -1,5 +1,4 @@
-  
-:root{
+:root {
   box-sizing: border-box;
   margin: 0;
   padding: 0;
@@ -10,7 +9,9 @@
   --body-bg: #212121;
   --card-bg: #424242;
 }
-*, ::before, ::after{
+*,
+::before,
+::after {
   box-sizing: border-box;
   margin: 0px;
   padding: 0px;
@@ -20,15 +21,16 @@ html {
   height: 100%;
 }
 
-body{
+body {
   background-color: white;
   height: 100%;
 }
 
 #root {
   height: 100%;
-  font-family: "Poopins" -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
-  background-color: #121212;
+  font-family: "Poopins" -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto,
+    Oxygen, Ubuntu, Cantarell, "Open Sans", "Helvetica Neue", sans-serif;
+  /* background-color: #121212; */
 }
 
 a {
@@ -42,7 +44,7 @@ a {
   display: grid;
   grid-template-columns: 1fr;
   grid-template-rows: 70px calc(100% - 70px);
-  background-color: #121212;
+  /* background-color: #121212; */
 }
 
 .App > .header {
@@ -67,8 +69,8 @@ a {
 
 /* Hide scrollbar for IE, Edge and Firefox */
 .App > .app_body {
-  -ms-overflow-style: none;  /* IE and Edge */
-  scrollbar-width: none;  /* Firefox */
+  -ms-overflow-style: none; /* IE and Edge */
+  scrollbar-width: none; /* Firefox */
 }
 
 .App > .app_body > .content {
