+++ src/index.js
@@ -1,6 +1,5 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
 import {BrowserRouter as Router} from 'react-router-dom'
@@ -16,7 +15,7 @@ ReactDOM.render(
           <App />
         </ChakraProvider>
       </Provider>
-    </Router>
+    </Router> 
   </React.StrictMode>,
   document.getElementById('root')
 );
