+++ src/pages/GiveHelp/GiveHelp.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const GiveHelp = () => {
+    return (
+        <div>
+            <h1>GiveHelp</h1>
+        </div>
+    )
+}
+
+export default GiveHelp
