+++ src/pages/Organization/Organization.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const Organization = () => {
+    return (
+        <div>
+            <h1>Organization</h1>
+        </div>
+    )
+}
+
+export default Organization
